fix(db): trip_id is of type bigint

+++ database/order_tbl.sql
@@ -14,7 +14,7 @@ CREATE TABLE order_segment_tbl
 (
     id               serial primary key,
     itinerary_id     int     not null references order_itinerary_tbl (id),
-    trip_id          int     not null references trip (id),
+    trip_id          bigint  not null references trip (id),
     start_station_id char(3) not null references stop (id),
     end_station_id   char(3) not null references stop (id),
     sequence         int     not null
